add gain to rtl_fm command
Adding gain value improves a lot the number of targets and decreases the time getting the first boats. Running rtl_test it is possible to get the supported gain values. 49.6 is the maximum of my dongle (RTL2832U, Rafael Micro R820T tuner), I don't know if this value can be valid for all.